Lazy load date pickers
