+++ App.js
@@ -1,7 +1,5 @@
 import {CssBaseline} from '@mui/material';
 import {ThemeProvider as MuiProvider} from '@mui/material/styles';
-import {AdapterDayjs} from '@mui/x-date-pickers/AdapterDayjs';
-import {LocalizationProvider as DateLocalizationProvider} from '@mui/x-date-pickers/LocalizationProvider';
 import {QueryClient, QueryClientProvider} from '@tanstack/react-query';
 import {ReactQueryDevtools} from '@tanstack/react-query-devtools';
 import Navigation from './src/Navigation';
@@ -26,12 +24,10 @@ export default function App() {
       <TokenLoadBuffer>
         <MuiProvider theme={theme}>
           <CssBaseline />
-          <DateLocalizationProvider dateAdapter={AdapterDayjs}>
-            <QueryClientProvider client={queryClient}>
-              {__DEV__ && <ReactQueryDevtools initialIsOpen={false} />}
-              <Navigation />
-            </QueryClientProvider>
-          </DateLocalizationProvider>
+          <QueryClientProvider client={queryClient}>
+            {__DEV__ && <ReactQueryDevtools initialIsOpen={false} />}
+            <Navigation />
+          </QueryClientProvider>
         </MuiProvider>
       </TokenLoadBuffer>
     </TokenProvider>
